Update Plugin.cs
Fixing build issues related to datatype

+++ JobRoulettePlugin/Plugin.cs
@@ -219,7 +219,7 @@ public sealed class Plugin : IDalamudPlugin
             return false;
         }
 
-        return playerState->ClassJobLevels[classJobId] > 0;
+        return playerState->ClassJobLevels[(int)classJobId] > 0;
     }
 }
